requirements.txt
scipy
pymupdf
PyPDF2
pytesseract
pillow
scipy
PyQt5
